Adds insert_test_blanks() RStudio addin
Adds an RStudio addin, `insert_test_blanks()`, that inserts a
templated testthat block into the current document. This test block
specifically checks for the presence of scaffolding blanks (`____`)
in the `learner_code` and fails if any are found. Also introduces a
helper function, `test_blanks_template()`, to return the template as a
string.